added file structure and hello world
