+++ .gitignore
@@ -15,3 +15,4 @@ Cargo.lock
 
 CSC362.code-workspace
 .obsidian
+sudoku-tui
